// src/app/model/dicom.model.ts
export enum DicomDecoded {
  patientId = 'x00100020',
  patientName = 'x00100010',
  patientBirthDate = 'x00100030',
  patientSex = 'x00100040',
  studyId = 'x00200010',
  studyDate = 'x00080020',
  studyTime = 'x00080030',
  studyDescription = 'x00081030',
  protocolName = 'x00181030',
  accessionNumber = 'x00080050',
  seriesNumber = 'x00200011',
  seriesDate = 'x00080021',
  seriesTime = 'x00080031',
  seriesDescription = 'x0008103e',
  modality = 'x00080060',
  bodyPartExamined = 'x00180015',
  instanceNumber = 'x00200013',
  acquisitionNumber = 'x00200012',
  acquisitionDate = 'x00080022',
  acquisitionTime = 'x00080032',
  contentDate = 'x00080023',
  contentTime = 'x00080033',
  manufacturer = 'x00080070',
  modelName = 'x00081090',
  stationName = 'x00081010',
  applicationEntityTitle = 'x00020016',
  institutionName = 'x00080080',
  softwareVersion = 'x00181020',
  implementationVersionName = 'x00020013',
  studyInstanceUID = 'x0020000d',
  seriesInstanceUID = 'x0020000e',
}